Added migration to rename profiles_tags pivot table to profile_tag Added model factory for Tag model Fixed: Namespace for App\Models\Tag
